// ecommerce-webapp-angular/src/app/model/data-model.ts
export class Food {
  id = 0;
  name = '';
  description = '';
  amount = 0;
  unit = '';
  calories = 0;
  macros: Macros;
}

export class Macros {
  protein = 0;
  fat = 0;
  hc = 0;
}
export class Persona {
  id: number;
  name: string;
  email: string;
  birthday: Date;
  height: number;
  weight: number;
  gender: number;
  country: number;
  postalCode: string;
  units: Units;
  goals: Goals;

}

export class Units {
  weight = '';
  height = '';
  distance = '';
  Energy = '';
  water = '';
}
export class Goals {
  initialWeight = 0;
  currentWeight = 0;
  desireWeight = 0;
  weeklyGoal = 0;
  activityLevel = 0;
  stepsAtDay = 0;
}

export class DietDays {

  date: Date = new Date();
  goalDay: number;
  totalCalories: number;
  exercise: number;
  steps: number;
  weight: number;
  meals: Meal[] = [];
}

export class Meal {
  name = '';
  foods: Food[] = [];
}